useAPIData.js moved to src
